requirements.txt
# requirements.txt
flask==2.0.1
gunicorn==20.1.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
Werkzeug==2.0.1